move prop type check custom function into utilities folder
